fix(ContactUs):  COM-2152
– update folders structure
– update styles
