+++ components/contact-us/ContactUs/ContactOptions/ContactOptions.styles.js
@@ -1,12 +1,7 @@
 import { css } from '@emotion/core'
 import { calcRem } from '../../../../utils/style/calcRem'
 
-const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
-  & {
-    display: flex;
-    flex-direction: column
-  }
-
+const base = ({ breakpoints: { tablet, mobile, desktop, above } }) => css`
   .subheading {
     font-family: Roboto, sans-serif;
     font-weight: 300;
@@ -14,12 +9,8 @@ const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
     line-height: ${calcRem(24)};
   }
 
-  .form {
-    position: relative;
-  }
-
   .separator {
-    position: absolute;
+    position: relative;
     display: flex;
     align-items: center;
     bottom: 0;
@@ -53,91 +44,19 @@ const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
     padding-bottom: ${calcRem(10)};
   }
 
-  .policy {
-    margin-top: ${calcRem(21)};
-    font-family: Roboto, sans-serif;
-    font-style: normal;
-    font-weight: 300;
-    font-size: 12px;
-    line-height: 16px;
-    color: #18191b;
-
-    > a {
-      color: #0076ff;
-    }
+  .separator,
+  .book-a-call {
+    margin-top: auto;
   }
 
-  .submit {
-    margin-top: ${calcRem(24)};
-
-    > button {
-      height: ${calcRem(40)};
-    }
-  }
-
-  .submit-text {
-    font-family: Roboto;
-    font-weight: 900;
-    font-size: ${calcRem(12)};
-    line-height: ${calcRem(24)};
-    text-align: center;
-    letter-spacing: 1.6px;
-    text-transform: uppercase;
-    color: #ffffff;
-  }
 
   ${desktop.all} {
-    & {
-
-    }
-
     .heading {
       margin-top: 99px;
     }
 
     .subheading {
       margin-top: 53px;
-    }
-
-    .form {
-      margin-top: 19px;
-      width: 320px;
-    }
-
-    .policy {
-      padding-right: 15px;
-    }
-  }
-
-  ${desktop.s} {
-    & {
-      padding-left: 120px;
-    }
-
-    .form {
-      width: 296px;
-    }
-
-    .policy {
-      padding-right: 0;
-    }
-
-    .separator {
-      right: -63px;
-    }
-  }
-
-  ${desktop.m} {
-    & {
-      padding-left: 128px;
-    }
-
-    .subheading {
-      margin-top: 53px;
-    }
-
-    .separator {
-      right: -63px;
     }
   }
 
@@ -151,13 +70,49 @@ const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
     }
 
     .separator {
-      right: -104px;
+      margin-left: 86px;
+    }
+
+    .book-a-call {
+      margin-left: 45px
+    }
+  }
+
+  ${desktop.m} {
+    & {
+      padding-left: 108px;
+    }
+
+    .subheading {
+      margin-top: 53px;
+    }
+
+    .separator {
+      margin-left: 45px;
+    }
+
+    .book-a-call {
+      margin-left: 6px
+    }
+  }
+
+  ${desktop.s} {
+    & {
+      padding-left: 100px;
+    }
+
+    .separator {
+      margin-left: 40px;
+    }
+
+    .book-a-call {
+      margin-left: 6px;
     }
   }
 
   ${tablet.all} {
     & {
-      padding-left: ${calcRem(64)};
+      padding-left: ${calcRem(57)};
     }
 
     .heading {
@@ -169,19 +124,19 @@ const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
       margin-top: ${calcRem(48)};
     }
 
-    .form {
-      margin-top: ${calcRem(19)};
-      width: ${calcRem(224)};
-    }
-
-    .policy {
-      font-size: ${calcRem(10)};
-      line-height: ${calcRem(16)};
-    }
-
     .separator {
-      right: ${calcRem(-48)};
+      margin-left: ${calcRem(30)};
       height: ${calcRem(316)};
+    }
+
+    .book-a-call {
+      margin-left: ${calcRem(-8)};
+    }
+  }
+
+  ${above.mobile} {
+    .form-wrapper {
+      display: flex;
     }
   }
 
@@ -199,17 +154,12 @@ const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
       margin-top: ${calcRem(32)};
     }
 
-    .form {
-      margin-top: ${calcRem(20)};
-      padding-bottom: ${calcRem(79)};
+    .form-wrapper {
+      display: block;
     }
 
-    .policy {
-      font-size: ${calcRem(10)};
-      line-height: ${calcRem(16)};
-    }
-
-    .separator {
+    .separator,
+    .book-a-call {
       display: none;
     }
   }
