+++ components/contact-us/ContactUs/ContactUs.styles.js
@@ -1,7 +1,6 @@
 import { css } from '@emotion/core'
-import calcRem from '../../../utils/style/calcRem'
 
-const base = ({ breakpoints: { tablet, mobile, desktop, above } }) => css`
+const base = ({ breakpoints: { tablet, mobile, desktop } }) => css`
   & {
     display: flex;
   }
@@ -10,7 +9,6 @@ const base = ({ breakpoints: { tablet, mobile, desktop, above } }) => css`
     top: 0;
     left: 0;
     width: 41.25%;
-    height: 100vh;
   }
 
   ${desktop.all} {
@@ -33,26 +31,10 @@ const base = ({ breakpoints: { tablet, mobile, desktop, above } }) => css`
     }
   }
 
-  ${above.mobile} {
-    .form-wrap {
-      display: flex
-    }
-
-    .book-a-call {
-      margin-top: ${calcRem(379)}; /* TODO: Тупой хак, что бы просто показать вёрстку заказчику, удалить потом */
-      margin-left: ${calcRem(72)};
-      margin-right: ${calcRem(20)};
-    }
-  }
-
   ${mobile.all} {
     & {
       margin-top: 3.5rem;
       flex-direction: column;
-    }
-
-    .book-a-call {
-      display: none;
     }
 
     .map-wrap {
